Move transformations source to a separate directory
