+++ include/pbrt/core/transformations/Quaternion.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "core.hpp"
-#include "math/Vector3.hpp"
+#include "pbrt/core/core.hpp"
+#include "pbrt/core/math/Vector3.hpp"
 
 namespace idragnev::pbrt {
     class Quaternion
