+++ include/pbrt/core/transformations/TRS.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "math/Vector3.hpp"
-#include "math/Matrix4x4.hpp"
+#include "pbrt/core/math/Vector3.hpp"
+#include "pbrt/core/math/Matrix4x4.hpp"
 #include "Quaternion.hpp"
 
 namespace idragnev::pbrt {
