+++ src/core/TransformedPrimitive.cpp
@@ -1,5 +1,5 @@
 #include "pbrt/core/TransformedPrimitive.hpp"
-#include "pbrt/core/Transformation.hpp"
+#include "pbrt/core/transformations/Transformation.hpp"
 #include "pbrt/core/SurfaceInteraction.hpp"
 #include "pbrt/core/Bounds3.hpp"
