+++ src/core/transformations/AnimatedTransformation.cpp
@@ -1,6 +1,6 @@
-#include "pbrt/core/AnimatedTransformation.hpp"
-#include "pbrt/core/Transformation.hpp"
-#include "pbrt/core/TRS.hpp"
+#include "pbrt/core/transformations/AnimatedTransformation.hpp"
+#include "pbrt/core/transformations/Transformation.hpp"
+#include "pbrt/core/transformations/TRS.hpp"
 #include "pbrt/core/Ray.hpp"
 #include "pbrt/core/RayDifferential.hpp"
 #include "pbrt/core/Bounds3.hpp"
