+++ src/shapes/Disk.cpp
@@ -1,5 +1,5 @@
 #include "pbrt/shapes/Disk.hpp"
-#include "pbrt/core/Transformation.hpp"
+#include "pbrt/core/transformations/Transformation.hpp"
 #include "pbrt/core/EFloat.hpp"
 #include "pbrt/core/Bounds3.hpp"
 #include "pbrt/core/math/Math.hpp"
