+++ src/shapes/Sphere.cpp
@@ -1,6 +1,6 @@
 #include "pbrt/shapes/Sphere.hpp"
 #include "pbrt/core/Bounds3.hpp"
-#include "pbrt/core/Transformation.hpp"
+#include "pbrt/core/transformations/Transformation.hpp"
 #include "pbrt/core/EFloat.hpp"
 #include "pbrt/core/math/Math.hpp"
